# src/knowledge_base/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Knowledge Base System API
  description: |
    # Unified Knowledge Base System API
    
    This API provides access to the Unified Knowledge Base System, allowing you to:
    
    - Add, update, and retrieve documents
    - Query the knowledge base with natural language
    - Manage users and API keys
    - Monitor system health and performance
    - Perform administrative tasks
    
    ## Authentication
    
    Most endpoints require authentication using either:
    - API key (via `X-API-Key` header)
    - Bearer token (via `Authorization` header)
    
    ## WebSocket Support
    
    Real-time communication is available through WebSocket at `/ws/{client_id}`.
    
    ## Rate Limiting
    
    API requests are subject to rate limiting based on your account tier.
  version: 1.0.0
  termsOfService: https://example.com/terms/
  contact:
    name: API Support
    url: https://example.com/support
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /admin/users:
    get:
      summary: List users
      description: List all users
      operationId: listUsers
      tags:
        - User Management
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      summary: Create user
      description: Create a new user
      operationId: createUser
      tags:
        - User Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/users/{user_id}:
    get:
      summary: Get user
      description: Get a user by ID
      operationId: getUser
      tags:
        - User Management
      parameters:
        - name: user_id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      summary: Update user
      description: Update a user
      operationId: updateUser
      tags:
        - User Management
      parameters:
        - name: user_id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      summary: Delete user
      description: Delete a user
      operationId: deleteUser
      tags:
        - User Management
      parameters:
        - name: user_id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/api-keys:
    get:
      summary: List API keys
      description: List API keys
      operationId: listApiKeys
      tags:
        - User Management
      parameters:
        - name: user_id
          in: query
          description: Filter by user ID
          schema:
            type: string
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKeyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      summary: Create API key
      description: Create a new API key
      operationId: createApiKey
      tags:
        - User Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/APIKeyCreate'
      responses:
        '200':
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKeyResponse'
        '400':
          description: Invalid API key data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/api-keys/{api_key_id}:
    put:
      summary: Update API key
      description: Update an API key
      operationId: updateApiKey
      tags:
        - User Management
      parameters:
        - name: api_key_id
          in: path
          description: API key ID
          required: true
          schema:
            type: string
        - name: name
          in: query
          description: New API key name
          schema:
            type: string
        - name: enabled
          in: query
          description: Whether the API key is enabled
          schema:
            type: boolean
        - name: expires_in_days
          in: query
          description: Days until expiration
          schema:
            type: integer
      responses:
        '200':
          description: API key updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKeyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      summary: Delete API key
      description: Delete an API key
      operationId: deleteApiKey
      tags:
        - User Management
      parameters:
        - name: api_key_id
          in: path
          description: API key ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    UserCreate:
      type: object
      properties:
        username:
          type: string
          description: Username
        email:
          type: string
          format: email
          description: Email address
        full_name:
          type: string
          description: Full name
        roles:
          type: array
          items:
            type: string
            enum: [reader, contributor, admin, system]
          description: User roles
          default: [reader]
        custom_permissions:
          type: array
          items:
            type: string
          description: Custom permissions
      required:
        - username
    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Email address
        full_name:
          type: string
          description: Full name
        disabled:
          type: boolean
          description: Whether the user is disabled
        roles:
          type: array
          items:
            type: string
            enum: [reader, contributor, admin, system]
          description: User roles
        custom_permissions:
          type: array
          items:
            type: string
          description: Custom permissions
    UserResponse:
      type: object
      properties:
        id:
          type: string
          description: User ID
        username:
          type: string
          description: Username
        email:
          type: string
          format: email
          description: Email address
        full_name:
          type: string
          description: Full name
        disabled:
          type: boolean
          description: Whether the user is disabled
        roles:
          type: array
          items:
            type: string
            enum: [reader, contributor, admin, system]
          description: User roles
        custom_permissions:
          type: array
          items:
            type: string
          description: Custom permissions
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - username
        - disabled
        - roles
        - custom_permissions
        - created_at
        - updated_at
    APIKeyCreate:
      type: object
      properties:
        name:
          type: string
          description: API key name
        user_id:
          type: string
          description: User ID
        expires_at:
          type: string
          format: date-time
          description: Expiration timestamp
      required:
        - name
        - user_id
    APIKeyResponse:
      type: object
      properties:
        id:
          type: string
          description: API key ID
        key:
          type: string
          description: API key value
        name:
          type: string
          description: API key name
        user_id:
          type: string
          description: User ID
        enabled:
          type: boolean
          description: Whether the API key is enabled
        expires_at:
          type: string
          format: date-time
          description: Expiration timestamp
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - id
        - key
        - name
        - user_id
        - enabled
        - created_at
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        error_type:
          type: string
          description: Error type
        status_code:
          type: integer
          description: HTTP status code
      required:
        - error
        - error_type
        - status_code
...
  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []